+++ demo/CUDA_Test/funset.hpp
@@ -1,6 +1,7 @@
 #ifndef FBC_CUDA_TEST_FUNSET_HPP_
 #define FBC_CUDA_TEST_FUNSET_HPP_
 
+#include <cstdlib>
 #include <vector>
 
 int test_image_reverse();
@@ -23,8 +24,8 @@ int streams_cpu(const int* a, const int* b, int* c, int length, float* elapsed_t
 int streams_gpu(const int* a, const int* b, int* c, int length, float* elapsed_time);
 
 int test_calculate_histogram();
-int calculate_histogram_cpu(const unsigned char* data, int length, size_t* hist, size_t& value, float* elapsed_time);
-int calculate_histogram_gpu(const unsigned char* data, int length, size_t* hist, size_t& value, float* elapsed_time);
+int calculate_histogram_cpu(const unsigned char* data, int length, unsigned int* hist, unsigned int& value, float* elapsed_time);
+int calculate_histogram_gpu(const unsigned char* data, int length, unsigned int * hist, unsigned int& value, float* elapsed_time);
 
 int test_heat_conduction();
 int heat_conduction_cpu(unsigned char* ptr, int width, int height, const float* src, float speed, float* elapsed_time);
